// apps/admin/src/api/generated/schemas/addonListResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * x-ear API
 * Backend API for x-ear CRM
 * OpenAPI spec version: 1.0.0
 */
import type { AddonListResponseData } from './addonListResponseData';
import type { AddonListResponseMessage } from './addonListResponseMessage';
import type { AddonListResponseError } from './addonListResponseError';
import type { AddonListResponseMeta } from './addonListResponseMeta';
import type { AddonListResponseRequestId } from './addonListResponseRequestId';

export interface AddonListResponse {
  success?: boolean;
  data?: AddonListResponseData;
  message?: AddonListResponseMessage;
  error?: AddonListResponseError;
  meta?: AddonListResponseMeta;
  requestId?: AddonListResponseRequestId;
  timestamp?: string;
}